% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{pbmc_h3k4me3_hg38}
\alias{pbmc_h3k4me3_hg38}
\title{pbmc_h3k4me3_hg38}
\format{
GRanges object
}
\usage{
data(pbmc_h3k4me3_hg38)
}
\description{
A set of ChIP-seq peaks from ENCODE, downloaded from AnnotationHub
in hg19 and lifted over to hg38. See vignette for details.
}
\keyword{datasets}
